rename: 공통 컴포넌트 Tab 폴더 변경
