+++ src/app/components/tab/TabButton.module.scss
@@ -2,7 +2,7 @@
 @use '@styles/variables' as Styled;
 
 /* variable */
-$styled-font-large: map.get(Styled.$typo, Label, Large);
+$styled-font-large: map.get(Styled.$typo, Title, Small);
 $styled-font-medium: map.get(Styled.$typo, Label, Medium);
 
 $secondary-100: map-get(Styled.$colors, Alert, Secondary100);
@@ -19,7 +19,7 @@ $secondary-900: map-get(Styled.$colors, Alert, Secondary900);
   font-weight: 600;
 
   &.active {
-    color: #15191e; // TODO 컬러파레트로 변경
+    color: $secondary-900;
 
     &::after {
       content: '';
@@ -42,7 +42,7 @@ $secondary-900: map-get(Styled.$colors, Alert, Secondary900);
   border-radius: 0.4rem;
 
   font: $styled-font-medium;
-  color: #15191e; // TODO 컬러파레트로 변경
+  color: $secondary-900;
   font-weight: 600;
 
   &.active {
